Organize game systems
