+++ .gitignore
@@ -6,6 +6,7 @@
 *.dylib
 sword
 game
+!game/
 
 *.tmx~
 *.tsx~
